feat: move SkillsMiddleware to core deepagents package
Closes #506

- Move SkillsMiddleware from deepagents-cli to core deepagents package
- Add re-exports in CLI for backward compatibility
- Add 17 unit tests for skills functionality
