+++ libs/deepagents-cli/deepagents_cli/skills/__init__.py
@@ -2,20 +2,29 @@
 
 Public API:
 - SkillsMiddleware: Middleware for integrating skills into agent execution
+- SkillMetadata: TypedDict for skill metadata
+- list_skills: Function to list skills from directories
 - execute_skills_command: Execute skills subcommands (list/create/info)
 - setup_skills_parser: Setup argparse configuration for skills commands
 
-All other components are internal implementation details.
+Note: SkillsMiddleware, SkillMetadata, and list_skills are re-exported from the
+core deepagents package for backward compatibility. New code should import
+directly from deepagents.middleware.skills.
 """
 
+# Re-export from core package for backward compatibility
+from deepagents.middleware.skills import SkillMetadata, SkillsMiddleware, list_skills
+
+# CLI-specific commands
 from deepagents_cli.skills.commands import (
     execute_skills_command,
     setup_skills_parser,
 )
-from deepagents_cli.skills.middleware import SkillsMiddleware
 
 __all__ = [
+    "SkillMetadata",
     "SkillsMiddleware",
     "execute_skills_command",
+    "list_skills",
     "setup_skills_parser",
 ]
