+++ libs/deepagents/deepagents/middleware/__init__.py
@@ -1,11 +1,13 @@
 """Middleware for the DeepAgent."""
 
 from deepagents.middleware.filesystem import FilesystemMiddleware
+from deepagents.middleware.skills import SkillsMiddleware
 from deepagents.middleware.subagents import CompiledSubAgent, SubAgent, SubAgentMiddleware
 
 __all__ = [
     "CompiledSubAgent",
     "FilesystemMiddleware",
+    "SkillsMiddleware",
     "SubAgent",
     "SubAgentMiddleware",
 ]
